refactor: chain configs. Consolidate into one location (#3059)
* Refactor configs to folders

- Move all configs to folders with ts files instead of json.
- Move pool config into configs folder instead of that big pools file.
- Move keys out into their own separate config and replace keys using
  typescript instead of that replace function.

* Switch all usage of POOLS constant to use config service

* Fixed issues with keys. Replaced all references to POOLS

* Add pools from #3053 and #3054

* Create new pools config and utils to simplify using network pool info

* Code cleanup and fixes

* Add ankMatic from #3058

* Revert moving pools to a config file, keep as POOLS from constants

* Add types to all config objects

* Remove strategies are they are unused

* Fix broken query file. Use poolMetadata function

* Add polygon pools from #3058

* Add goerli pool changes from #3043

* Remove strategies from arbitrum config

* Fix config type reference

* Fix another reference

* Remove appAddresses as it doesn't appear to be used anywhere
