+++ src/components/contextual/pages/pool/PoolContractDetails.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { POOLS } from '@/constants/pools';
+import { poolMetadata } from '@/composables/usePool';
 import { shortenLabel } from '@/lib/utils';
 import { Pool, PoolType } from '@/services/pool/types';
 import useWeb3 from '@/services/web3/useWeb3';
@@ -59,7 +60,7 @@ const data = computed(() => {
     },
     {
       title: t('poolName'),
-      value: POOLS.Metadata[id]?.name || name,
+      value: poolMetadata(id)?.name || name,
     },
     {
       title: t('poolSymbol'),
