+++ src/components/contextual/pages/pool/PoolMigrationCard/NewPoolData.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import { Pool } from '@/services/pool/types';
 import {
+  hasIcon,
   orderedPoolTokens,
   orderedTokenAddresses,
 } from '@/composables/usePool';
-import { POOLS } from '@/constants/pools';
 import TokenPills from '@/components/tables/PoolsTable/TokenPills/TokenPills.vue';
 import BalChipNew from '@/components/chips/BalChipNew.vue';
 import useNetwork from '@/composables/useNetwork';
@@ -22,9 +22,7 @@ type Props = {
 defineProps<Props>();
 
 function iconAddresses(pool: Pool) {
-  return POOLS.Metadata[pool.id]?.hasIcon
-    ? [pool.address]
-    : orderedTokenAddresses(pool);
+  return hasIcon(pool.id) ? [pool.address] : orderedTokenAddresses(pool);
 }
 
 /**
