+++ src/composables/usePool.ts
@@ -4,13 +4,8 @@ import { computed, Ref } from 'vue';
 
 import { POOL_MIGRATIONS } from '@/components/forms/pool_actions/MigrateForm/constants';
 import { ALLOWED_RATE_PROVIDERS } from '@/constants/rateProviders';
-import {
-  POOLS,
-  APR_THRESHOLD,
-  DeprecatedDetails,
-  PoolMetadata,
-  POOLS_MAP,
-} from '@/constants/pools';
+import { APR_THRESHOLD } from '@/constants/pools';
+import { DeprecatedDetails, PoolMetadata } from '@/types/pools';
 import {
   bnum,
   includesAddress,
@@ -19,6 +14,7 @@ import {
 } from '@/lib/utils';
 import { includesWstEth } from '@/lib/utils/balancer/lido';
 import { configService } from '@/services/config/config.service';
+import configs from '@/lib/config';
 
 import {
   isTestnet,
@@ -33,11 +29,17 @@ import { AnyPool, Pool, PoolToken, SubPool } from '@/services/pool/types';
 import { hasBalEmissions } from '@/services/staking/utils';
 import { uniq, uniqWith, cloneDeep } from 'lodash';
 
+const POOLS = configService.network.pools;
+
 /**
  * METHODS
  */
 export function addressFor(poolId: string): string {
   return getAddress(poolId.slice(0, 42));
+}
+
+export function hasIcon(poolId: string): boolean {
+  return !!poolMetadata(poolId)?.hasIcon;
 }
 
 export function isLinear(poolType: PoolType): boolean {
@@ -589,7 +591,7 @@ export function poolMetadata(
   id: string,
   network = networkId.value
 ): PoolMetadata | undefined {
-  return POOLS_MAP[network]?.Metadata[id.toLowerCase()];
+  return configs[network]?.pools.Metadata[id.toLowerCase()];
 }
 
 /**
