--- src/lib/config/arbitrum.json
@@ -1,92 +1,0 @@
-{
-  "key": "42161",
-  "chainId": 42161,
-  "chainName": "Arbitrum",
-  "name": "Arbitrum",
-  "shortName": "Arbitrum",
-  "slug": "arbitrum",
-  "network": "arbitrum-one",
-  "unknown": false,
-  "rpc": "https://arb-mainnet.g.alchemy.com/v2/{{ALCHEMY_KEY}}",
-  "ws": "wss://arb-mainnet.g.alchemy.com/v2/{{ALCHEMY_KEY}}",
-  "publicRpc": "https://arb1.arbitrum.io/rpc",
-  "explorer": "https://arbiscan.io",
-  "explorerName": "Arbiscan",
-  "subgraph": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-arbitrum-v2",
-  "balancerApi": "https://api.balancer.fi",
-  "poolsUrlV1": "",
-  "poolsUrlV2": "",
-  "subgraphs": {
-    "main": [
-      "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-arbitrum-v2"
-    ],
-    "aave": "",
-    "gauge": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gauges-arbitrum",
-    "blocks": "https://api.thegraph.com/subgraphs/name/ianlapham/arbitrum-one-blocks"
-  },
-  "supportsEIP1559": false,
-  "supportsElementPools": false,
-  "blockTime": 2,
-  "nativeAsset": {
-    "name": "Ether",
-    "address": "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
-    "symbol": "ETH",
-    "decimals": 18,
-    "deeplinkId": "ether",
-    "logoURI": "tokens/eth.png",
-    "minTransactionBuffer": "0.05"
-  },
-  "addresses": {
-    "merkleRedeem": "0x6bd0B17713aaa29A2d7c9A39dDc120114f9fD809",
-    "merkleOrchard": "0x751A0bC0e3f75b38e01Cf25bFCE7fF36DE1C87DE",
-    "merkleOrchardV2": "0x9805dcfD25e6De36bad8fe9D3Fe2c9b44B764102",
-    "multicall": "0x80C7DD17B01855a6D2347444a0FCC36136a314de",
-    "authorizer": "0x6207ed574152496c9B072C24FD87cE9cd9E17320",
-    "vault": "0xBA12222222228d8Ba445958a75a0704d566BF2C8",
-    "weightedPoolFactory": "0xf1665E19bc105BE4EDD3739F88315cC699cc5b65",
-    "stablePoolFactory": "0x2433477A10FC5d31B9513C638F19eE85CaED53Fd",
-    "weth": "0x82aF49447D8a07e3bd95BD0d56f35241523fBab1",
-    "stMATIC": "",
-    "rETH": "0xEC70Dcb4A1EFa46b8F2D97C310C9c4790ba5ffA8",
-    "stETH": "",
-    "wstETH": "0x5979d7b546e38e414f7e9822514be443a4800529",
-    "lidoRelayer": "",
-    "balancerHelpers": "0x77d46184d22CA6a3726a2F500c776767b6A3d6Ab",
-    "batchRelayer": "0x466262c2a275aB106E54D95B5B04603e12b58cA1",
-    "batchRelayerV4": "0x5bf3B7c14b10f16939d63Bd679264A1Aa951B4D5",
-    "gaugeFactory": "0xb08e16cfc07c684daa2f93c70323badb2a6cbfd2",
-    "balancerMinter": "",
-    "gaugeController": "",
-    "tokenAdmin": "",
-    "veBAL": "",
-    "veDelegationProxy": "",
-    "veBALHelpers": "",
-    "feeDistributor": "",
-    "feeDistributorDeprecated": "",
-    "faucet": "",
-    "gaugeRewardsHelper": "0xA0DAbEBAAd1b243BBb243f933013d560819eB66f"
-  },
-  "keys": {
-    "infura": "daaa68ec242643719749dd1caba2fc66",
-    "alchemy": "VBeQgTCRqqPtuuEPsFzRdwKXzDyN6aFh",
-    "balancerApi": "da2-7a3ukmnw7bexndpx5x522uafui"
-  },
-  "strategies": {
-    "0": {
-      "type": "0",
-      "name": "stablePool"
-    },
-    "1": {
-      "type": "1",
-      "name": "weightedPool"
-    },
-    "2": {
-      "type": "2",
-      "name": "weightedPool"
-    }
-  },
-  "gauges": {
-    "type": 3,
-    "weight": 0
-  }
-}
