+++ src/lib/config/arbitrum/contracts.ts
@@ -1,0 +1,34 @@
+import { Contracts } from '../types';
+
+const contracts: Contracts = {
+  merkleRedeem: '0x6bd0B17713aaa29A2d7c9A39dDc120114f9fD809',
+  merkleOrchard: '0x751A0bC0e3f75b38e01Cf25bFCE7fF36DE1C87DE',
+  merkleOrchardV2: '0x9805dcfD25e6De36bad8fe9D3Fe2c9b44B764102',
+  multicall: '0x80C7DD17B01855a6D2347444a0FCC36136a314de',
+  authorizer: '0x6207ed574152496c9B072C24FD87cE9cd9E17320',
+  vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+  weightedPoolFactory: '0xf1665E19bc105BE4EDD3739F88315cC699cc5b65',
+  stablePoolFactory: '0x2433477A10FC5d31B9513C638F19eE85CaED53Fd',
+  weth: '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
+  stMATIC: '',
+  rETH: '0xEC70Dcb4A1EFa46b8F2D97C310C9c4790ba5ffA8',
+  stETH: '',
+  wstETH: '0x5979d7b546e38e414f7e9822514be443a4800529',
+  lidoRelayer: '',
+  balancerHelpers: '0x77d46184d22CA6a3726a2F500c776767b6A3d6Ab',
+  batchRelayer: '0x466262c2a275aB106E54D95B5B04603e12b58cA1',
+  batchRelayerV4: '0x5bf3B7c14b10f16939d63Bd679264A1Aa951B4D5',
+  gaugeFactory: '0xb08e16cfc07c684daa2f93c70323badb2a6cbfd2',
+  balancerMinter: '',
+  gaugeController: '',
+  tokenAdmin: '',
+  veBAL: '',
+  veDelegationProxy: '',
+  veBALHelpers: '',
+  feeDistributor: '',
+  feeDistributorDeprecated: '',
+  faucet: '',
+  gaugeRewardsHelper: '0xA0DAbEBAAd1b243BBb243f933013d560819eB66f',
+};
+
+export default contracts;
