+++ src/lib/config/docker/contracts.ts
@@ -1,0 +1,33 @@
+import { Contracts } from '../types';
+
+const contracts: Contracts = {
+  merkleRedeem: '',
+  merkleOrchard: '',
+  merkleOrchardV2: '',
+  multicall: '0xfe17B701A8F44D9D18D4f83dB71Bf9442D0003f1',
+  authorizer: '0x6Bff77238256fFF717e5Cc9A96640dC0D068b14D',
+  vault: '0x293f6758Fe4a4343E367Ac8E6A697eC6315fBEd7',
+  weightedPoolFactory: '0xE48Aae4178612fdF77af25C83fa1099491A8AD1B',
+  stablePoolFactory: '0xb4F97F452D71C3bba3c30c554305A7250c08b23A',
+  weth: '0x6c37Fa9557C135dF4323004D3f0549f3e94F112D',
+  rETH: '0xae78736Cd615f374D3085123A210448E74Fc6393',
+  stMATIC: '',
+  stETH: '0xae7ab96520de3a18e5e111b5eaab095312d7fe84',
+  wstETH: '0x7f39c581f595b53c5cb19bd0b3f8da6c935e2ca0',
+  lidoRelayer: '',
+  balancerHelpers: '',
+  batchRelayer: '',
+  batchRelayerV4: '',
+  veBAL: '',
+  gaugeController: '',
+  gaugeFactory: '',
+  balancerMinter: '',
+  tokenAdmin: '',
+  veDelegationProxy: '',
+  veBALHelpers: '',
+  feeDistributor: '',
+  feeDistributorDeprecated: '',
+  faucet: '',
+};
+
+export default contracts;
