+++ src/lib/config/global.ts
@@ -1,0 +1,4 @@
+export const keys = {
+  infura: 'daaa68ec242643719749dd1caba2fc66',
+  alchemy: 'VBeQgTCRqqPtuuEPsFzRdwKXzDyN6aFh',
+};
