--- src/lib/config/gnosis-chain.json
@@ -1,90 +1,0 @@
-{
-  "key": "100",
-  "chainId": 100,
-  "chainName": "Gnosis Chain",
-  "name": "Gnosis Chain",
-  "shortName": "Gnosis",
-  "slug": "gnosis-chain",
-  "network": "gnosis-chain",
-  "unknown": false,
-  "rpc": "https://poa-xdai.gateway.pokt.network/v1/lb/91bc0e12a76e7a84dd76189d",
-  "ws": "wss://rpc.gnosischain.com/wss",
-  "publicRpc": "https://rpc.gnosis.gateway.fm",
-  "loggingRpc": "https://rpc.gnosis.gateway.fm",
-  "explorer": "https://gnosisscan.io",
-  "explorerName": "Gnosisscan",
-  "balancerApi": "https://api.balancer.fi",
-  "subgraph": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gnosis-chain-v2",
-  "poolsUrlV2": "",
-  "subgraphs": {
-    "main": [
-      "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gnosis-chain-v2"
-    ],
-    "aave": "",
-    "gauge": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gauges-gnosis-chain-b",
-    "blocks": ""
-  },
-  "supportsEIP1559": true,
-  "supportsElementPools": true,
-  "blockTime": 5,
-  "nativeAsset": {
-    "name": "xDAI",
-    "address": "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
-    "symbol": "xDAI",
-    "decimals": 18,
-    "deeplinkId": "xdai",
-    "logoURI": "tokens/xdai.png",
-    "minTransactionBuffer": "0.05"
-  },
-  "addresses": {
-    "merkleRedeem": "",
-    "merkleOrchard": "0x35b50C7955e7D0E8298e043C9F4dCDef737b9f5a",
-    "multicall": "0xbb6fab6b627947dae0a75808250d8b2652952cb5",
-    "authorizer": "0xA331D84eC860Bf466b4CdCcFb4aC09a1B43F3aE6",
-    "vault": "0xBA12222222228d8Ba445958a75a0704d566BF2C8",
-    "weightedPoolFactory": "0xC128a9954e6c874eA3d62ce62B468bA073093F25",
-    "stablePoolFactory": "0xf23b4DB826DbA14c0e857029dfF076b1c0264843",
-    "weth": "0xe91d153e0b41518a2ce8dd3d7944fa863463a97d",
-    "rETH": "",
-    "stMATIC": "",
-    "stETH": "",
-    "wstETH": "",
-    "lidoRelayer": "",
-    "balancerHelpers": "0x8E9aa87E45e92bad84D5F8DD1bff34Fb92637dE9",
-    "batchRelayer": "0x7b9B6f094DC2Bd1c12024b0D9CC63d6993Be1888",
-    "batchRelayerV4": "0xeF606F58A4FD0fCcb066c6203d0994694d3eB2D3",
-    "gaugeFactory": "0x809B79b53F18E9bc08A961ED4678B901aC93213a",
-    "balancerMinter": "",
-    "gaugeController": "",
-    "tokenAdmin": "",
-    "veBAL": "",
-    "veDelegationProxy": "",
-    "veBALHelpers": "",
-    "feeDistributor": "",
-    "feeDistributorDeprecated": "",
-    "faucet": "",
-    "gaugeRewardsHelper": "0xf7D5DcE55E6D47852F054697BAB6A1B48A00ddbd"
-  },
-  "keys": {
-    "infura": "daaa68ec242643719749dd1caba2fc66",
-    "alchemy": "oGLgncMVHNWltnK0nTfiryxQ6gYemKQO"
-  },
-  "strategies": {
-    "0": {
-      "type": "0",
-      "name": "stablePool"
-    },
-    "1": {
-      "type": "1",
-      "name": "weightedPool"
-    },
-    "2": {
-      "type": "2",
-      "name": "weightedPool"
-    }
-  },
-  "gauges": {
-    "type": 2,
-    "weight": 100
-  }
-}
