+++ src/lib/config/gnosis-chain/contracts.ts
@@ -1,0 +1,33 @@
+import { Contracts } from '../types';
+
+const contracts: Contracts = {
+  merkleRedeem: '',
+  merkleOrchard: '0x35b50C7955e7D0E8298e043C9F4dCDef737b9f5a',
+  multicall: '0xbb6fab6b627947dae0a75808250d8b2652952cb5',
+  authorizer: '0xA331D84eC860Bf466b4CdCcFb4aC09a1B43F3aE6',
+  vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+  weightedPoolFactory: '0xC128a9954e6c874eA3d62ce62B468bA073093F25',
+  stablePoolFactory: '0xf23b4DB826DbA14c0e857029dfF076b1c0264843',
+  weth: '0xe91d153e0b41518a2ce8dd3d7944fa863463a97d',
+  rETH: '',
+  stMATIC: '',
+  stETH: '',
+  wstETH: '',
+  lidoRelayer: '',
+  balancerHelpers: '0x8E9aa87E45e92bad84D5F8DD1bff34Fb92637dE9',
+  batchRelayer: '0x7b9B6f094DC2Bd1c12024b0D9CC63d6993Be1888',
+  batchRelayerV4: '0xeF606F58A4FD0fCcb066c6203d0994694d3eB2D3',
+  gaugeFactory: '0x809B79b53F18E9bc08A961ED4678B901aC93213a',
+  balancerMinter: '',
+  gaugeController: '',
+  tokenAdmin: '',
+  veBAL: '',
+  veDelegationProxy: '',
+  veBALHelpers: '',
+  feeDistributor: '',
+  feeDistributorDeprecated: '',
+  faucet: '',
+  gaugeRewardsHelper: '0xf7D5DcE55E6D47852F054697BAB6A1B48A00ddbd',
+};
+
+export default contracts;
