+++ src/lib/config/goerli/keys.ts
@@ -1,0 +1,8 @@
+import { Keys } from '../types';
+
+const keys: Keys = {
+  infura: 'daaa68ec242643719749dd1caba2fc66',
+  alchemy: 'oGLgncMVHNWltnK0nTfiryxQ6gYemKQO',
+};
+
+export default keys;
