+++ src/lib/config/mainnet/keys.ts
@@ -1,0 +1,10 @@
+import { Keys } from '../types';
+
+const keys: Keys = {
+  infura: 'daaa68ec242643719749dd1caba2fc66',
+  alchemy: 'cQGZUiTLRCFsQS7kbRxPJK4eH4fTTu88',
+  graph: '7c7b4c36244b0b86171a3931eaf9bb23',
+  balancerApi: 'da2-7a3ukmnw7bexndpx5x522uafui',
+};
+
+export default keys;
