+++ src/lib/config/optimism/contracts.ts
@@ -1,0 +1,32 @@
+import { Contracts } from '../types';
+
+const contracts: Contracts = {
+  merkleRedeem: '',
+  merkleOrchard: '',
+  multicall: '',
+  authorizer: '',
+  vault: '',
+  weightedPoolFactory: '0xA0DAbEBAAd1b243BBb243f933013d560819eB66f',
+  stablePoolFactory: '',
+  weth: '',
+  rETH: '',
+  stMATIC: '',
+  stETH: '',
+  wstETH: '',
+  lidoRelayer: '',
+  balancerHelpers: '',
+  batchRelayer: '',
+  batchRelayerV4: '0x1a58897Ab366082028ced3740900ecBD765Af738',
+  gaugeFactory: '',
+  balancerMinter: '',
+  gaugeController: '',
+  tokenAdmin: '',
+  veBAL: '',
+  veDelegationProxy: '',
+  veBALHelpers: '',
+  feeDistributor: '',
+  feeDistributorDeprecated: '',
+  faucet: '',
+};
+
+export default contracts;
