--- src/lib/config/polygon.json
@@ -1,91 +1,0 @@
-{
-  "key": "137",
-  "chainId": 137,
-  "chainName": "Polygon",
-  "name": "Polygon Mainnet",
-  "shortName": "Polygon",
-  "slug": "polygon",
-  "network": "polygon",
-  "unknown": false,
-  "rpc": "https://polygon-mainnet.infura.io/v3/{{INFURA_KEY}}",
-  "ws": "wss://polygon-mainnet.g.alchemy.com/v2/{{ALCHEMY_KEY}}",
-  "publicRpc": "https://polygon-rpc.com",
-  "explorer": "https://polygonscan.com",
-  "explorerName": "Polygonscan",
-  "subgraph": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-polygon-prune-v2",
-  "balancerApi": "https://api.balancer.fi",
-  "poolsUrlV2": "",
-  "subgraphs": {
-    "main": [
-      "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-polygon-prune-v2"
-    ],
-    "aave": "https://api.thegraph.com/subgraphs/name/aave/aave-v2-matic",
-    "gauge": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gauges-polygon",
-    "blocks": "https://api.thegraph.com/subgraphs/name/ianlapham/polygon-blocks"
-  },
-  "supportsEIP1559": true,
-  "supportsElementPools": false,
-  "blockTime": 4,
-  "nativeAsset": {
-    "name": "Matic",
-    "address": "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
-    "symbol": "MATIC",
-    "decimals": 18,
-    "deeplinkId": "matic",
-    "logoURI": "tokens/matic.svg",
-    "minTransactionBuffer": "0.1"
-  },
-  "addresses": {
-    "merkleRedeem": "",
-    "merkleOrchard": "0x0F3e0c4218b7b0108a3643cFe9D3ec0d4F57c54e",
-    "merkleOrchardV2": "0xc3ccacE87f6d3A81724075ADcb5ddd85a8A1bB68",
-    "multicall": "0x275617327c958bD06b5D6b871E7f491D76113dd8",
-    "authorizer": "0xa331d84ec860bf466b4cdccfb4ac09a1b43f3ae6",
-    "vault": "0xBA12222222228d8Ba445958a75a0704d566BF2C8",
-    "weightedPoolFactory": "0x82e4cFaef85b1B6299935340c964C942280327f4",
-    "stablePoolFactory": "0xc66Ba2B6595D3613CCab350C886aCE23866EDe24",
-    "weth": "0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270",
-    "rETH": "0x0266F4F08D82372CF0FcbCCc0Ff74309089c74d1",
-    "stMATIC": "0x3a58a54c066fdc0f2d55fc9c89f0415c92ebf3c4",
-    "stETH": "",
-    "wstETH": "",
-    "lidoRelayer": "",
-    "balancerHelpers": "0x94905e703fead7f0fd0eee355d267ee909784e6d",
-    "batchRelayer": "0xcf6a66E32dCa0e26AcC3426b851FD8aCbF12Dac7",
-    "batchRelayerV4": "0x28A224d9d398a1eBB7BA69BCA515898966Bb1B6b",
-    "gaugeFactory": "0x3b8ca519122cdd8efb272b0d3085453404b25bd0",
-    "balancerMinter": "",
-    "gaugeController": "",
-    "tokenAdmin": "",
-    "veBAL": "",
-    "veDelegationProxy": "",
-    "veBALHelpers": "",
-    "feeDistributor": "",
-    "feeDistributorDeprecated": "",
-    "faucet": "",
-    "gaugeRewardsHelper": "0xaEb406b0E430BF5Ea2Dc0B9Fe62E4E53f74B3a33"
-  },
-  "keys": {
-    "infura": "daaa68ec242643719749dd1caba2fc66",
-    "alchemy": "ODJ9G5Ipv-Hb2zTWMNbUFIqv9WtqBOc2",
-    "balancerApi": "da2-7a3ukmnw7bexndpx5x522uafui"
-  },
-  "strategies": {
-    "0": {
-      "type": "0",
-      "name": "stablePool"
-    },
-    "1": {
-      "type": "1",
-      "name": "weightedPool"
-    },
-    "2": {
-      "type": "2",
-      "name": "weightedPool"
-    }
-  },
-  "gauges": {
-    "type": 4,
-    "weight": 0
-  }
-}
