+++ src/lib/config/polygon/contracts.ts
@@ -1,0 +1,34 @@
+import { Contracts } from '../types';
+
+const contracts: Contracts = {
+  merkleRedeem: '',
+  merkleOrchard: '0x0F3e0c4218b7b0108a3643cFe9D3ec0d4F57c54e',
+  merkleOrchardV2: '0xc3ccacE87f6d3A81724075ADcb5ddd85a8A1bB68',
+  multicall: '0x275617327c958bD06b5D6b871E7f491D76113dd8',
+  authorizer: '0xa331d84ec860bf466b4cdccfb4ac09a1b43f3ae6',
+  vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+  weightedPoolFactory: '0x82e4cFaef85b1B6299935340c964C942280327f4',
+  stablePoolFactory: '0xc66Ba2B6595D3613CCab350C886aCE23866EDe24',
+  weth: '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270',
+  rETH: '0x0266F4F08D82372CF0FcbCCc0Ff74309089c74d1',
+  stMATIC: '0x3a58a54c066fdc0f2d55fc9c89f0415c92ebf3c4',
+  stETH: '',
+  wstETH: '',
+  lidoRelayer: '',
+  balancerHelpers: '0x94905e703fead7f0fd0eee355d267ee909784e6d',
+  batchRelayer: '0xcf6a66E32dCa0e26AcC3426b851FD8aCbF12Dac7',
+  batchRelayerV4: '0x28A224d9d398a1eBB7BA69BCA515898966Bb1B6b',
+  gaugeFactory: '0x3b8ca519122cdd8efb272b0d3085453404b25bd0',
+  balancerMinter: '',
+  gaugeController: '',
+  tokenAdmin: '',
+  veBAL: '',
+  veDelegationProxy: '',
+  veBALHelpers: '',
+  feeDistributor: '',
+  feeDistributorDeprecated: '',
+  faucet: '',
+  gaugeRewardsHelper: '0xaEb406b0E430BF5Ea2Dc0B9Fe62E4E53f74B3a33',
+};
+
+export default contracts;
