+++ src/lib/config/polygon/keys.ts
@@ -1,0 +1,9 @@
+import { Keys } from '../types';
+
+const keys: Keys = {
+  infura: 'daaa68ec242643719749dd1caba2fc66',
+  alchemy: 'ODJ9G5Ipv-Hb2zTWMNbUFIqv9WtqBOc2',
+  balancerApi: 'da2-7a3ukmnw7bexndpx5x522uafui',
+};
+
+export default keys;
