+++ src/lib/config/polygon/pools.ts
@@ -1,0 +1,179 @@
+import { Pools } from '@/types/pools';
+
+const pools: Pools = {
+  IdsMap: {
+    xMatic: {
+      v1: '0xc17636e36398602dd37bb5d1b3a9008c7629005f0002000000000000000004c4',
+      v2: '0xb20fc01d21a50d2c734c4a1262b4404d41fa7bf000000000000000000000075c',
+    },
+    stMatic: {
+      v1: '0xaf5e0b5425de1f5a630a8cb5aa9d97b8141c908d000200000000000000000366',
+      v2: '0x8159462d255c1d24915cb51ec361f700174cd99400000000000000000000075d',
+    },
+    mai4: {
+      mai4: '0x06df3b2bbb68adc8b0e302443692037ed9f91b42000000000000000000000012',
+      maiBbaUsd:
+        '0xb54b2125b711cd183edd3dd09433439d5396165200000000000000000000075e',
+    },
+  },
+  Pagination: {
+    PerPage: 10,
+    PerPool: 10,
+    PerPoolInitial: 5,
+  },
+  DelegateOwner: '0xba1ba1ba1ba1ba1ba1ba1ba1ba1ba1ba1ba1ba1b',
+  ZeroAddress: '0x0000000000000000000000000000000000000000',
+  DynamicFees: {
+    Gauntlet: [],
+  },
+  BlockList: [''],
+  IncludedPoolTypes: [
+    'Weighted',
+    'Stable',
+    'MetaStable',
+    'LiquidityBootstrapping',
+    'Investment',
+    'StablePhantom',
+    'ComposableStable',
+    'FX',
+  ],
+  Stable: {
+    AllowList: [
+      '0x06df3b2bbb68adc8b0e302443692037ed9f91b42000000000000000000000012', // polygon MAI/DAI/USDC/USDT
+      '0xfeadd389a5c427952d8fdb8057d6c8ba1156cc5600020000000000000000001e', // polygon WBTC/renBTC
+      '0xf38cf113d2d4f60c36cbd95af2f48a9a0167045a00000000000000000000005b', // polygon,
+      '0x0d34e5dd4d8f043557145598e4e2dc286b35fd4f000000000000000000000068', // tusd polygon
+      '0x5028497af0c9a54ea8c6d42a054c0341b9fc616800020000000000000000007b', // dusd polygon
+      '0xaf5e0b5425de1f5a630a8cb5aa9d97b8141c908d000200000000000000000366', // polygon staked matic
+      '0xc31a37105b94ab4efca1954a14f059af11fcd9bb000000000000000000000455', // 4pool
+      '0xc17636e36398602dd37bb5d1b3a9008c7629005f0002000000000000000004c4', // maticx metastable
+      '0xb4b22bd6cdad0ab828be6f8a4086dfa54e9b373600020000000000000000058f', // Polygon tetuBAL-80BAL-20WETH
+      '0xb797adfb7b268faeaa90cadbfed464c76ee599cd0002000000000000000005ba', // tetuBAL-80BAL-20WETH V2 (with short name)
+      '0x0b8319061732b34cab22445fa83b81f950e4b7ed000000000000000000000709',
+      '0xaf5e0b5425de1f5a630a8cb5aa9d97b8141c908d000200000000000000000366',
+      '0x8159462d255c1d24915cb51ec361f700174cd99400000000000000000000075d',
+      '0xb20fc01d21a50d2c734c4a1262b4404d41fa7bf000000000000000000000075c',
+      '0xb54b2125b711cd183edd3dd09433439d5396165200000000000000000000075e', // mai / bb-am-USD
+      '0x48e6b98ef6329f8f0a30ebb8c7c960330d64808500000000000000000000075b', // bb-am-USD
+      '0xa48d164f6eb0edc68bd03b56fa59e12f24499ad10000000000000000000007c4', // ageur stable
+      '0x2d46979fd4c5f7a04f65111399cff3da2dab5bd9000000000000000000000807', // ankr stable
+      '0x47401399b2eca91930c99126df20a11531f99465000000000000000000000840', // 3brl pool
+      '0x76afd126f46ab4fdf2ece8b1a2c149f7cf95d9fb00000000000000000000085c', // 2cad
+      '0x92bc61bd96f275ea5507a3e1e5fbf0bc69cc98dc00000000000000000000085d', // 2sgd
+      '0x7d60a4cb5ca92e2da965637025122296ea6854f900000000000000000000085e', // 2eur par
+      '0x94970a3f6a6aab442aefad68ff57abec0b9e3c0100000000000000000000085f', // 2eur eurt
+      '0x7913e4c8d00044689ff5c7c12d2f1b4a2fde4994000000000000000000000860', // 2eur eure
+      '0x7982c1b61abdc36942301ff2377d92b43784f120000000000000000000000861', // 2try
+      '0x7f408fbcfc88917bff6a79b0ed0646fa090627de000000000000000000000863', // 2jpy
+      '0x9e0a3a9b5a4e0b6dc299a56ef19002f23842be8d000000000000000000000862', // 2mxn
+      '0x05f21bacc4fd8590d1eaca9830a64b66a733316c00000000000000000000087e', // tetuQI
+      '0x02d2e2d7a89d6c5cb3681cfcb6f7dac02a55eda400000000000000000000088f', // csMatic
+      '0xe22483774bd8611be2ad2f4194078dac9159f4ba0000000000000000000008f0', // 2BRL
+      '0xbf29ef6e23af0ac5b6bf931c8b3f1080f5bc120600000000000000000000091f', // vQi stable
+      '0x34a81e8956bf20b7448b31990a2c06f96830a6e4000200000000000000000a14', // wUSDR
+      '0x5dee84ffa2dc27419ba7b3419d7146e53e4f7ded000200000000000000000a4e', // frxETH / WETH
+      '0xd80ef9fabfdc3b52e17f74c383cf88ee2efbf0b6000000000000000000000a65', // tetu boosted
+      '0x513cdee00251f39de280d9e5f771a6eafebcc88e000000000000000000000a6b', // 2eur/par
+      '0x77e97d4908be63394bc5dff72c8c7bddf1699882000000000000000000000a6a', // augeur
+      '0x3db543faf7a92052de7860c5c9debabee59ed5bd000000000000000000000a62', // 4usd
+      '0x65fe9314be50890fb01457be076fafd05ff32b9a000000000000000000000a96', // wsteth/eth
+      '0xb3d658d5b95bf04e2932370dd1ff976fe18dd66a000000000000000000000ace', // bb-t-USD (tetu managed boosted pool)
+      '0x9a020bdc2faff5bd24c6acc2020d01ff9f2c627a000000000000000000000ae2', // overnight davos usd
+      '0x19017f2919a5fb7eca1f0d142330644dc2045423000000000000000000000af9', // 2EUR (EURe)
+      '0x02559a4fa0f3dae55820a65eb48b7a2fcd82f361000000000000000000000af8', // 2EUR (EURs)
+      '0x36a0ee903841584f47e3c774b59e0cbfba46080f000000000000000000000b0a', // ankMatic/Matic
+      '0x10b040038f87219d9b42e025e3bd9b8095c87dd9000000000000000000000b11', // bb-t-MATIC (tetu managed boosted pool)
+    ],
+  },
+  Investment: {
+    AllowList: [''],
+  },
+  Factories: {
+    '0xa5bf2ddf098bb0ef6d120c98217dd6b141c74ee0': 'oracleWeightedPool',
+    '0x8e9aa87e45e92bad84d5f8dd1bff34fb92637de9': 'weightedPool',
+    '0xc66ba2b6595d3613ccab350c886ace23866ede24': 'stablePool',
+    '0xdae7e32adc5d490a43ccba1f0c736033f2b4efca': 'stablePool', // Metastable
+    '0x751a0bc0e3f75b38e01cf25bfce7ff36de1c87de': 'liquidityBootstrappingPool', // LBP
+    '0x41b953164995c11c81da73d212ed8af25741b7ac': 'liquidityBootstrappingPool', // LBP (zero protocol fee)
+    '0x0f7bb7ce7b6ed9366f9b6b910adefe72dc538193': 'managedPool', // Polygon Managed
+    '0xc128a9954e6c874ea3d62ce62b468ba073093f25': 'boostedPool', // polygon stablephantom
+    '0xca96c4f198d343e251b1a01f3eba061ef3da73c1': 'stablePool', // stable pool v2,
+    '0x136fd06fa01ecf624c7f2b3cb15742c1339dc2c4': 'composableStablePool', // ComposableStable
+    '0x0e39c3d9b2ec765efd9c5c70bb290b1fcd8536e3': 'weightedPool', // weighted pool v2
+    '0x7bc6c0e73edaa66ef3f6e2f27b0ee8661834c6c9': 'composableStablePool', // ComposableStable V3
+    '0x82e4cfaef85b1b6299935340c964c942280327f4': 'weightedPool', // weighted pool v3
+    '0x627d759314d5c4007b461a74ebafa7ebc5dfed71': 'fx', // fx
+  },
+  Stakable: {
+    AllowList: [
+      '0x0297e37f1873d2dab4487aa67cd56b58e2f27875000100000000000000000002',
+      '0x03cd191f589d12b0582a99808cf19851e468e6b500010000000000000000000a',
+      '0x06df3b2bbb68adc8b0e302443692037ed9f91b42000000000000000000000012',
+      '0x0d34e5dd4d8f043557145598e4e2dc286b35fd4f000000000000000000000068',
+      '0x10f21c9bd8128a29aa785ab2de0d044dcdd79436000200000000000000000059',
+      '0x186084ff790c65088ba694df11758fae4943ee9e000200000000000000000032',
+      '0x36128d5436d2d70cab39c9af9cce146c38554ff0000100000000000000000008',
+      '0x5a6ae1fd70d04ba4a279fc219dfabc53825cb01d00020000000000000000020e',
+      '0x614b5038611729ed49e0ded154d8a5d3af9d1d9e00010000000000000000001d',
+      '0x7c9cf12d783821d5c63d8e9427af5c44bad92445000100000000000000000051',
+      '0x805ca3ccc61cc231851dee2da6aabff0a7714aa7000200000000000000000361',
+      '0xaf5e0b5425de1f5a630a8cb5aa9d97b8141c908d000200000000000000000366',
+      '0xb204bf10bc3a5435017d3db247f56da601dfe08a0002000000000000000000fe',
+      '0xc31a37105b94ab4efca1954a14f059af11fcd9bb000000000000000000000455',
+      '0xce66904b68f1f070332cbc631de7ee98b650b499000100000000000000000009',
+      '0xcf354603a9aebd2ff9f33e1b04246d8ea204ae9500020000000000000000005a',
+      '0xdb1db6e248d7bb4175f6e5a382d0a03fe3dcc813000100000000000000000035',
+      '0xea4e073c8ac859f2994c07e627178719c8002dc00002000000000000000003dc',
+      '0xfeadd389a5c427952d8fdb8057d6c8ba1156cc5600020000000000000000001e',
+      '0xc17636e36398602dd37bb5d1b3a9008c7629005f0002000000000000000004c4',
+      '0x2dbc9ab0160087ae59474fb7bed95b9e808fa6bc0001000000000000000003db',
+      '0xb797adfb7b268faeaa90cadbfed464c76ee599cd0002000000000000000005ba',
+      '0x8f9dd2064eb38e8e40f2ab67bde27c0e16ea9b080002000000000000000004ca',
+      '0x48e6b98ef6329f8f0a30ebb8c7c960330d64808500000000000000000000075b',
+      '0xb54b2125b711cd183edd3dd09433439d5396165200000000000000000000075e',
+      '0x8159462d255c1d24915cb51ec361f700174cd99400000000000000000000075d',
+      '0xb20fc01d21a50d2c734c4a1262b4404d41fa7bf000000000000000000000075c',
+      '0x8ac5fafe2e52e52f5352aec64b64ff8b305e1d4a0002000000000000000007ab',
+      '0x05f21bacc4fd8590d1eaca9830a64b66a733316c00000000000000000000087e',
+      '0x4973f591784d9c94052a6c3ebd553fcd37bb0e5500020000000000000000087f',
+      '0xe2f706ef1f7240b803aae877c9c762644bb808d80002000000000000000008c2',
+      '0x4a0b73f0d13ff6d43e304a174697e3d5cfd310a400020000000000000000091c',
+      '0xe22483774bd8611be2ad2f4194078dac9159f4ba0000000000000000000008f0',
+      '0xa48d164f6eb0edc68bd03b56fa59e12f24499ad10000000000000000000007c4',
+      '0x7d60a4cb5ca92e2da965637025122296ea6854f900000000000000000000085e',
+      '0x34a81e8956bf20b7448b31990a2c06f96830a6e4000200000000000000000a14',
+      '0xf3312968c7d768c19107731100ece7d4780b47b2000200000000000000000a50',
+      '0x5dee84ffa2dc27419ba7b3419d7146e53e4f7ded000200000000000000000a4e',
+      '0xeab6455f8a99390b941a33bbdaf615abdf93455e000200000000000000000a66',
+      '0x577f6076e558818a5df21ce4acde9a9623ec0b4c000200000000000000000a64',
+      '0x77e97d4908be63394bc5dff72c8c7bddf1699882000000000000000000000a6a',
+      '0x513cdee00251f39de280d9e5f771a6eafebcc88e000000000000000000000a6b',
+      '0xd80ef9fabfdc3b52e17f74c383cf88ee2efbf0b6000000000000000000000a65',
+      '0x65fe9314be50890fb01457be076fafd05ff32b9a000000000000000000000a96',
+    ],
+  },
+  Metadata: {
+    '0x48e6b98ef6329f8f0a30ebb8c7c960330d64808500000000000000000000075b': {
+      name: 'Balancer Boosted Aave USD (Polygon)',
+      hasIcon: true,
+    },
+    '0xb3d658d5b95bf04e2932370dd1ff976fe18dd66a000000000000000000000ace': {
+      name: 'Balancer Boosted Tetu USD (Polygon)',
+      hasIcon: true,
+    },
+    '0x10b040038f87219d9b42e025e3bd9b8095c87dd9000000000000000000000b11': {
+      name: 'Balancer Boosted Tetu MATIC (Polygon)',
+      hasIcon: true,
+    },
+  },
+  DisabledJoins: [
+    '0xfeadd389a5c427952d8fdb8057d6c8ba1156cc5600020000000000000000001e',
+    '0xb54b2125b711cd183edd3dd09433439d5396165200000000000000000000075e',
+  ],
+  BrandedRedirect: {
+    '0x726e324c29a1e49309672b244bdc4ff62a270407000200000000000000000702':
+      'xave',
+  },
+};
+
+export default pools;
