+++ src/lib/config/test/keys.ts
@@ -1,0 +1,8 @@
+import { Keys } from '../types';
+
+const keys: Keys = {
+  infura: 'daaa68ec242643719749dd1caba2fc66',
+  alchemy: 'cQGZUiTLRCFsQS7kbRxPJK4eH4fTTu88',
+};
+
+export default keys;
